Data/archive/requirements.txt
Flask==2.2.2
pandas==1.5.0
google-cloud-bigquery==3.3.5
plotly==5.14.1
scikit-learn==1.2.2
gunicorn==20.1.0
pyarrow==9.0.0
db-dtypes
google-cloud-bigquery-storage
